Transpile tests without type checking
