+++ .mocharc.client.js
@@ -2,16 +2,28 @@
 
 const base = require('./.mocharc.base.json');
 
+/**
+ * Mocha will run `ts-node` without doing type checking to speed-up the tests. It should be fine as `npm run typecheck`
+ * covers test files too.
+ */
+
+Object.assign(process.env, {
+	TS_NODE_FILES: true,
+	TS_NODE_TRANSPILE_ONLY: true,
+}, process.env);
+
 module.exports = {
 	...base, // see https://github.com/mochajs/mocha/issues/3916
 	require: [
 		...base.require,
-		'./tests/mocks/client/register.ts',
+		'./tests/setup/registerWebApiMocks.ts',
 		'./tests/setup/cleanupTestingLibrary.ts',
 	],
+	exit: false,
+	parallel: true,
+	slow: 200,
 	spec: [
 		'client/**/*.spec.ts',
 		'client/**/*.spec.tsx',
 	],
-	exit: true,
 };
