+++ tests/setup/registerWebApiMocks.ts
@@ -1,0 +1,5 @@
+import { enableBlobUrlsMock } from '../mocks/client/blobUrls';
+import { enableJsdom } from '../mocks/client/jsdom';
+
+enableJsdom();
+enableBlobUrlsMock();
